+++ changelog.md
@@ -3,6 +3,10 @@
 ## NEXT
 
 - Added documentation for the `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` class.
+- Fixed `repr` bug in the `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` class.
+- Added equality method for the `cai_causal_graph.graph_components.TimeSeriesNode` class.
+- Extended unit tests for the `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph` and
+  `cai_causal_graph.graph_components.TimeSeriesNode` classes.
 
 ## 0.2.0